asr/requirements.txt
# Dependencies for the ASR container.

fastapi==0.115.12
uvicorn[standard]==0.34.2

numpy>=1.22.0
cuda-python>=12.3

noisereduce==3.0.0

pyspellchecker
spellwise
